requirements.txt
paho-mqtt==1.6.1
cryptography==41.0.7
